requirements-data.txt
openpyxl
python-benedict
python-fsutil
python-slugify
requests